# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RuangSadar API",
        "description": "Dokumentasi API untuk RuangSadar",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ruang-sadar-production.up.railway.app/",
            "description": "Production Server (Railway)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/journal": {
            "get": {
                "tags": [
                    "Journal"
                ],
                "summary": "Ambil riwayat jurnal pengguna",
                "description": "Menampilkan daftar jurnal khusus milik user yang sedang login",
                "operationId": "bead93448320c17eec49e73f8911af31",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data jurnal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Journal"
                ],
                "summary": "Simpan jurnal baru",
                "description": "Menambahkan catatan jurnal baru untuk user yang sedang login",
                "operationId": "9c8043e3a2a19ad4ff66313e22edea62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fokus Hari Ini"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Hari ini saya merasa lebih tenang dan produktif setelah meditasi pagi."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jurnal berhasil disimpan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/appointments": {
            "post": {
                "tags": [
                    "Patient"
                ],
                "summary": "Buat jadwal konsultasi baru",
                "description": "Menyimpan data dari form Atur Jadwal",
                "operationId": "4a053cfa62a42a38721cf5dd97e4c68b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "therapist_id",
                                    "full_name",
                                    "consultation_date",
                                    "topic",
                                    "method"
                                ],
                                "properties": {
                                    "therapist_id": {
                                        "description": "ID Psikolog/Terapis yang dipilih dari list",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "consultation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-25"
                                    },
                                    "topic": {
                                        "type": "string",
                                        "example": "Saya sering merasa cemas akhir-akhir ini karena tugas menumpuk."
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "online"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil dibuat"
                    },
                    "403": {
                        "description": "Akses ditolak (Bukan Pasien)"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/profile": {
            "put": {
                "tags": [
                    "Patient"
                ],
                "summary": "Ubah profil pasien",
                "description": "Mengupdate data nama dan username khusus untuk pasien yang sedang login",
                "operationId": "ce829cd417104f537f3d6c93fd87dd2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Christian Hadi Candra"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "christian_candra"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui"
                    },
                    "403": {
                        "description": "Akses ditolak (Bukan Pasien)"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/therapists/profile": {
            "put": {
                "tags": [
                    "Therapist"
                ],
                "summary": "Ubah profil terapis",
                "description": "Mengupdate data profil khusus untuk terapis yang sedang login",
                "operationId": "97310cf8cc0b8290585b0685a45d36cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Dian Sastro, M.Psi"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "dian_sastro_psi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui"
                    },
                    "403": {
                        "description": "Akses ditolak (Bukan Terapis)"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/therapists/appointments": {
            "get": {
                "tags": [
                    "Therapist"
                ],
                "summary": "Lihat daftar jadwal konsultasi",
                "description": "Mengambil semua jadwal pasien yang masuk ke terapis ini",
                "operationId": "f3fd07b2f32bec6224d2c99ef3d1c2d1",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil jadwal"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/therapists/appointments/{id}/status": {
            "patch": {
                "tags": [
                    "Therapist"
                ],
                "summary": "Ubah status jadwal konsultasi",
                "description": "Mengubah status jadwal menjadi approved, completed, atau cancelled",
                "operationId": "e4be8c7408fc222aa3b104183620a652",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Jadwal (Appointment)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Pilihan: approved, completed, cancelled",
                                        "type": "string",
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diubah"
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/videos": {
            "get": {
                "tags": [
                    "Video"
                ],
                "summary": "Ambil semua daftar video",
                "description": "Mendapatkan list video, bisa difilter berdasarkan kategori melalui query parameter",
                "operationId": "146978877bfb1a6b73cf56db141eaa3a",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter video berdasarkan kategori (opsional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Anxiety"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data video"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/videos/{id}": {
            "get": {
                "tags": [
                    "Video"
                ],
                "summary": "Ambil detail video berdasarkan ID",
                "description": "Mendapatkan data spesifik dari satu video",
                "operationId": "d605cccdd1603726ce4c710a466cbf0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Video",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menemukan video"
                    },
                    "404": {
                        "description": "Video tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Patient",
            "description": "API untuk fitur dan data khusus pasien"
        },
        {
            "name": "Therapist",
            "description": "API untuk fitur dan data khusus terapis"
        },
        {
            "name": "Video",
            "description": "API untuk mengelola video"
        },
        {
            "name": "Journal",
            "description": "Journal"
        }
    ]
}